Refactor of types and structs
