+++ multiaccounts/settings/enums.go
@@ -1,0 +1,17 @@
+package settings
+
+type ProfilePicturesVisibilityType int
+
+const (
+	ProfilePicturesVisibilityContactsOnly ProfilePicturesVisibilityType = iota + 1
+	ProfilePicturesVisibilityEveryone
+	ProfilePicturesVisibilityNone
+)
+
+type ProfilePicturesShowToType int
+
+const (
+	ProfilePicturesShowToContactsOnly ProfilePicturesShowToType = iota + 1
+	ProfilePicturesShowToEveryone
+	ProfilePicturesShowToNone
+)
