+++ multiaccounts/settings/structs.go
@@ -4,24 +4,88 @@ import (
 	"encoding/json"
 
 	"github.com/status-im/status-go/eth-node/types"
+	"github.com/status-im/status-go/protocol/common"
+	"github.com/status-im/status-go/protocol/protobuf"
 )
 
-type ProfilePicturesVisibilityType int
+type ValueHandler func(interface{}) (interface{}, error)
+type SyncSettingProtobufFactoryInterface func(interface{}, uint64, string) (*common.RawMessage, error)
+type SyncSettingProtobufFactoryStruct func(Settings, uint64, string) (*common.RawMessage, error)
+type SyncSettingProtobufToValue func(setting *protobuf.SyncSetting) interface{}
 
-const (
-	ProfilePicturesVisibilityContactsOnly ProfilePicturesVisibilityType = iota + 1
-	ProfilePicturesVisibilityEveryone
-	ProfilePicturesVisibilityNone
-)
+// SyncProtobufFactory represents a collection of functionality to generate and parse *protobuf.SyncSetting
+type SyncProtobufFactory struct {
+	inactive          bool
+	fromInterface     SyncSettingProtobufFactoryInterface
+	fromStruct        SyncSettingProtobufFactoryStruct
+	valueFromProtobuf SyncSettingProtobufToValue
+	protobufType      protobuf.SyncSetting_Type
+}
 
-type ProfilePicturesShowToType int
+func (spf *SyncProtobufFactory) Inactive() bool {
+	return spf.inactive
+}
 
-const (
-	ProfilePicturesShowToContactsOnly ProfilePicturesShowToType = iota + 1
-	ProfilePicturesShowToEveryone
-	ProfilePicturesShowToNone
-)
+func (spf *SyncProtobufFactory) FromInterface() SyncSettingProtobufFactoryInterface {
+	return spf.fromInterface
+}
 
+func (spf *SyncProtobufFactory) FromStruct() SyncSettingProtobufFactoryStruct {
+	return spf.fromStruct
+}
+
+func (spf *SyncProtobufFactory) ExtractValueFromProtobuf() SyncSettingProtobufToValue {
+	return spf.valueFromProtobuf
+}
+
+func (spf *SyncProtobufFactory) SyncSettingProtobufType() protobuf.SyncSetting_Type {
+	return spf.protobufType
+}
+
+// SyncSettingField represents a binding between a Value and a SettingField
+type SyncSettingField struct {
+	SettingField
+	Value interface{}
+}
+
+func (s SyncSettingField) MarshalJSON() ([]byte, error) {
+	alias := struct {
+		Name  string      `json:"name"`
+		Value interface{} `json:"value"`
+	}{
+		s.reactFieldName,
+		s.Value,
+	}
+
+	return json.Marshal(alias)
+}
+
+// SettingField represents an individual setting in the database, it contains context dependant names and optional
+// pre-store value parsing, along with optional *SyncProtobufFactory
+type SettingField struct {
+	reactFieldName      string
+	dBColumnName        string
+	valueHandler        ValueHandler
+	syncProtobufFactory *SyncProtobufFactory
+}
+
+func (s SettingField) GetReactName() string {
+	return s.reactFieldName
+}
+
+func (s SettingField) GetDBName() string {
+	return s.dBColumnName
+}
+
+func (s SettingField) ValueHandler() ValueHandler {
+	return s.valueHandler
+}
+
+func (s SettingField) SyncProtobufFactory() *SyncProtobufFactory {
+	return s.syncProtobufFactory
+}
+
+// Settings represents the entire setting row stored in the application db
 type Settings struct {
 	// required
 	Address                   types.Address    `json:"address"`
